+++ src/main/java/de/uni_passau/fim/se2/litterbox/ast/model/metadata/MonitorListMetadata.java
@@ -9,7 +9,7 @@ public class MonitorListMetadata extends MonitorMetadata {
     private List<String> values;
 
     public MonitorListMetadata(String id, String mode, String opcode, MonitorParamMetadataList paramsMetadata,
-                               String spriteName, int width, int height, int x, int y, boolean visible,
+                               String spriteName, double width, double height, double x, double y, boolean visible,
                                List<String> values) {
         super(id, mode, opcode, paramsMetadata, spriteName, width, height, x, y, visible);
         this.values = values;
