+++ src/main/java/de/uni_passau/fim/se2/litterbox/ast/model/metadata/TopBlockMetadata.java
@@ -4,22 +4,22 @@ import de.uni_passau.fim.se2.litterbox.ast.model.metadata.astLists.InputMetadata
 import de.uni_passau.fim.se2.litterbox.ast.visitor.ScratchVisitor;
 
 public class TopBlockMetadata extends BlockMetadata {
-    private int xPos;
-    private int yPos;
+    private double xPos;
+    private double yPos;
 
     public TopBlockMetadata(CommentMetadata commentMetadata, String blockId, String opcode, String next,
                             String parent, InputMetadataList inputMetadata, InputMetadataList fields,
-                            boolean topLevel, boolean shadow, int xPos, int yPos) {
+                            boolean topLevel, boolean shadow, double xPos, double yPos) {
         super(commentMetadata, blockId, opcode, next, parent, inputMetadata, fields, topLevel, shadow);
         this.xPos = xPos;
         this.yPos = yPos;
     }
 
-    public int getxPos() {
+    public double getxPos() {
         return xPos;
     }
 
-    public int getyPos() {
+    public double getyPos() {
         return yPos;
     }
